Custom `renderer` objects for modifying rendering of chart elements.
At this stage, this is just a proposal for an approach to handling complex customization without embedding that complexity within Squid.

The intention is to allow for users with complex rendering requirements — beyond Squid’s built-in options — to roll their own custom variations without having to modify or extend Squid itself.

So far, only regular (not stacked) columns (`.column`) and the boxes in legends (`.legend_box`) are supported. But this has been structured to be adapted to almost any element of the charts to allow for a high level of customizability without over-extending/customizing the Squid API.

Should this approach be adopted by the Squid core developers, it could eventually include methods to override the rendering found in almost any of the Squid::Plotter methods.

SECURITY CONSIDERATION: The `Squid::Configuration` class requires loading of settings from strings. For the existing settings, this has been fine because they have all been simple values (booleans, numbers, text). But for the Renderer setting, the string has to be evaluated as code. This presents a potential security issue if people are running their chart code with user-supplied settings values. If this code is adopted, we should both look at validation/restriction of the string being evaluated, and include an advisory in the documentation of the renderer setting.