[renderer] Move axis label rendering to custom `renderer` object.
Allows for the rendering of the label text and tick marks to be overridden by a custom renderer class/object.

This does not complete the transition of rendering functionality out of `Squid::Plotter` and into the `DefaultRenderer`. So far, only the column boxes, legend boxes, and now the axis labels, have been implemented.